# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/gateway_transactions:
    get:
      tags:
        - GatewayTransactions
      summary: List GatewayTransactions
      description: Get a paginated list of GatewayTransactions with optional filtering and sorting
      operationId: listGatewayTransactions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayTransactionList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - GatewayTransactions
      summary: Create GatewayTransaction
      description: Create a new GatewayTransaction
      operationId: createGatewayTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGatewayTransactionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayTransaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/payment_gateway_providers:
    get:
      tags:
        - PaymentGatewayProviders
      summary: List PaymentGatewayProviders
      description: Get a paginated list of PaymentGatewayProviders with optional filtering and sorting
      operationId: listPaymentGatewayProviders
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentGatewayProviderList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PaymentGatewayProviders
      summary: Create PaymentGatewayProvider
      description: Create a new PaymentGatewayProvider
      operationId: createPaymentGatewayProvider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentGatewayProviderRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentGatewayProvider'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    GatewayTransaction:
      type: object
      required:
        - id
        - company_id
        - provider_id
        - provider_code
        - provider_transaction_id
        - direction
        - gross_amount
        - fee_amount
        - net_amount
        - currency
        - status
        - posting_state
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        provider_id:
          type: string
          format: uuid
        provider_code:
          type: string
        provider_transaction_id:
          type: string
          maxLength: 140
        direction:
          type: string
        party_type:
          nullable: true
          type: string
        party_id:
          nullable: true
          type: string
          format: uuid
        gross_amount:
          type: string
        fee_amount:
          type: string
        net_amount:
          type: string
        currency:
          type: string
          maxLength: 3
        status:
          type: string
        posting_state:
          type: string
        payment_entry_id:
          nullable: true
          type: string
          format: uuid
        fee_post_id:
          nullable: true
          type: string
          format: uuid
        settled_at:
          nullable: true
          type: string
          format: date-time
        reference_no:
          nullable: true
          type: string
          maxLength: 140
        raw_payload:
          nullable: true
          type: object
        metadata:
          type: object
...
    CreateGatewayTransactionRequest:
      type: object
      required:
        - company_id
        - provider_id
        - provider_code
        - provider_transaction_id
        - direction
        - gross_amount
        - fee_amount
        - net_amount
        - currency
        - status
        - posting_state
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        provider_id:
          type: string
          format: uuid
        provider_code:
          type: string
        provider_transaction_id:
          type: string
          maxLength: 140
        direction:
          type: string
        party_type:
          nullable: true
          type: string
        party_id:
          nullable: true
          type: string
          format: uuid
        gross_amount:
          type: string
        fee_amount:
          type: string
        net_amount:
          type: string
        currency:
          type: string
          maxLength: 3
        status:
          type: string
        posting_state:
          type: string
        payment_entry_id:
          nullable: true
          type: string
          format: uuid
        fee_post_id:
          nullable: true
          type: string
          format: uuid
        settled_at:
          nullable: true
          type: string
          format: date-time
        reference_no:
          nullable: true
          type: string
          maxLength: 140
        raw_payload:
          nullable: true
          type: object
        metadata:
          type: object
...
    GatewayTransactionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GatewayTransaction'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaymentGatewayProvider:
      type: object
      required:
        - id
        - code
        - company_id
        - display_name
        - is_active
        - metadata
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        company_id:
          type: string
          format: uuid
        display_name:
          type: string
          maxLength: 120
        credentials_ref:
          nullable: true
          type: string
          maxLength: 255
        fee_account_id:
          nullable: true
          type: string
          format: uuid
        settlement_account_id:
          nullable: true
          type: string
          format: uuid
        is_active:
          type: boolean
        metadata:
          type: object
...
    CreatePaymentGatewayProviderRequest:
      type: object
      required:
        - code
        - company_id
        - display_name
        - is_active
        - metadata
      properties:
        code:
          type: string
        company_id:
          type: string
          format: uuid
        display_name:
          type: string
          maxLength: 120
        credentials_ref:
          nullable: true
          type: string
          maxLength: 255
        fee_account_id:
          nullable: true
          type: string
          format: uuid
        settlement_account_id:
          nullable: true
          type: string
          format: uuid
        is_active:
          type: boolean
        metadata:
          type: object
...
    PaymentGatewayProviderList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PaymentGatewayProvider'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []