# src/presentation/http/gateway_transaction_handler.rs
//! GatewayTransaction REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{DateTime, Utc};
use rust_decimal::Decimal;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{GatewayTransactionService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateGatewayTransactionDto, UpdateGatewayTransactionDto, PatchGatewayTransactionDto, GatewayTransactionResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum GatewayTransactionError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for GatewayTransactionError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for GatewayTransactionError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "GATEWAYTRANSACTION_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "GATEWAYTRANSACTION_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "GATEWAYTRANSACTION_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "GATEWAYTRANSACTION_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for GatewayTransaction.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /gateway_transactions | List with pagination |
/// | POST | /gateway_transactions | Create new |
/// | GET | /gateway_transactions/:id | Get by ID |
/// | PUT | /gateway_transactions/:id | Full update |
/// | PATCH | /gateway_transactions/:id | Partial update |
/// | DELETE | /gateway_transactions/:id | Soft delete |
/// | POST | /gateway_transactions/bulk | Bulk create |
/// | POST | /gateway_transactions/upsert | Upsert |
/// | GET | /gateway_transactions/trash | List deleted |
/// | POST | /gateway_transactions/:id/restore | Restore |
/// | DELETE | /gateway_transactions/empty | Empty trash |
/// | GET | /gateway_transactions/:id/deleted | Get deleted by ID |
/// | DELETE | /gateway_transactions/trash/:id | Permanent delete from trash |
/// | GET | /gateway_transactions/count | Count active entities |
/// | GET | /gateway_transactions/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(GatewayTransactionService::with_repository(repository));
/// let router = create_gateway_transaction_routes(service);
/// ```
pub fn create_gateway_transaction_routes(service: Arc<GatewayTransactionService>) -> Router {
    BackboneCrudHandler::<GatewayTransactionService, GatewayTransaction, CreateGatewayTransactionDto, UpdateGatewayTransactionDto, GatewayTransactionResponseDto>::routes(
        service,
        "/gateway_transactions",
    )
}

/// Create Axum router with only the read (GET) endpoints for GatewayTransaction.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_gateway_transaction_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_gateway_transaction_read_routes(service: Arc<GatewayTransactionService>) -> Router {
    BackboneCrudHandler::<GatewayTransactionService, GatewayTransaction, CreateGatewayTransactionDto, UpdateGatewayTransactionDto, GatewayTransactionResponseDto>::read_routes(
        service,
        "/gateway_transactions",
    )
}

/// Create Axum router with only the write (mutation) endpoints for GatewayTransaction.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
pub fn create_gateway_transaction_write_routes(service: Arc<GatewayTransactionService>) -> Router {
    BackboneCrudHandler::<GatewayTransactionService, GatewayTransaction, CreateGatewayTransactionDto, UpdateGatewayTransactionDto, GatewayTransactionResponseDto>::write_routes(
        service,
        "/gateway_transactions",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_gateway_transaction_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<GatewayTransactionService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_gateway_transaction_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}