# src/presentation/http/payment_gateway_provider_handler.rs
//! PaymentGatewayProvider REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{PaymentGatewayProviderService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreatePaymentGatewayProviderDto, UpdatePaymentGatewayProviderDto, PatchPaymentGatewayProviderDto, PaymentGatewayProviderResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum PaymentGatewayProviderError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for PaymentGatewayProviderError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for PaymentGatewayProviderError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "PAYMENTGATEWAYPROVIDER_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "PAYMENTGATEWAYPROVIDER_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "PAYMENTGATEWAYPROVIDER_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "PAYMENTGATEWAYPROVIDER_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for PaymentGatewayProvider.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /payment_gateway_providers | List with pagination |
/// | POST | /payment_gateway_providers | Create new |
/// | GET | /payment_gateway_providers/:id | Get by ID |
/// | PUT | /payment_gateway_providers/:id | Full update |
/// | PATCH | /payment_gateway_providers/:id | Partial update |
/// | DELETE | /payment_gateway_providers/:id | Soft delete |
/// | POST | /payment_gateway_providers/bulk | Bulk create |
/// | POST | /payment_gateway_providers/upsert | Upsert |
/// | GET | /payment_gateway_providers/trash | List deleted |
/// | POST | /payment_gateway_providers/:id/restore | Restore |
/// | DELETE | /payment_gateway_providers/empty | Empty trash |
/// | GET | /payment_gateway_providers/:id/deleted | Get deleted by ID |
/// | DELETE | /payment_gateway_providers/trash/:id | Permanent delete from trash |
/// | GET | /payment_gateway_providers/count | Count active entities |
/// | GET | /payment_gateway_providers/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(PaymentGatewayProviderService::with_repository(repository));
/// let router = create_payment_gateway_provider_routes(service);
/// ```
pub fn create_payment_gateway_provider_routes(service: Arc<PaymentGatewayProviderService>) -> Router {
    BackboneCrudHandler::<PaymentGatewayProviderService, PaymentGatewayProvider, CreatePaymentGatewayProviderDto, UpdatePaymentGatewayProviderDto, PaymentGatewayProviderResponseDto>::routes(
        service,
        "/payment_gateway_providers",
    )
}

/// Create Axum router with only the read (GET) endpoints for PaymentGatewayProvider.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_payment_gateway_provider_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_payment_gateway_provider_read_routes(service: Arc<PaymentGatewayProviderService>) -> Router {
    BackboneCrudHandler::<PaymentGatewayProviderService, PaymentGatewayProvider, CreatePaymentGatewayProviderDto, UpdatePaymentGatewayProviderDto, PaymentGatewayProviderResponseDto>::read_routes(
        service,
        "/payment_gateway_providers",
    )
}

/// Create Axum router with only the write (mutation) endpoints for PaymentGatewayProvider.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
pub fn create_payment_gateway_provider_write_routes(service: Arc<PaymentGatewayProviderService>) -> Router {
    BackboneCrudHandler::<PaymentGatewayProviderService, PaymentGatewayProvider, CreatePaymentGatewayProviderDto, UpdatePaymentGatewayProviderDto, PaymentGatewayProviderResponseDto>::write_routes(
        service,
        "/payment_gateway_providers",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_payment_gateway_provider_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<PaymentGatewayProviderService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_payment_gateway_provider_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}